+++ src/actions/refreshPage.js
@@ -1,5 +1,8 @@
 import setCurrentNavigationByRequestAddress from './navigation/setCurrentNavigationByRequestAddress';
+import loadUserSettingsFromStorage from "./settings/loadUserSettingsFromStorage";
 
-export default () => (dispatch) => {
+export default () => (dispatch) => new Promise((resolve) => {
     setCurrentNavigationByRequestAddress()(dispatch);
-};
+    loadUserSettingsFromStorage()(dispatch);
+    resolve();
+});