+++ src/actions/welcomeUser.js
@@ -1,6 +1,7 @@
 import getCurrentUserId from './user/getCurrentUserId';
 import initializeUserIdsFromStorage from './user/initializeUserIdsFromStorage';
 import storageStub from '../storage/storageStub';
+import loadUserSettingsFromRepository from './settings/loadUserSettingsFromRepository';
 import loadWishlistItems from "./wishlist/loadWishlistItems";
 
 const storage = sessionStorage || storageStub;
@@ -11,6 +12,7 @@ export default () => (dispatch, getState) => new Promise((resolve) => {
     initializeUserIdsFromStorage()(dispatch);
     getCurrentUserId()(dispatch, getState).then((currentUserId) => {
         loadWishlistItems(currentUserId)(dispatch);
+        loadUserSettingsFromRepository(currentUserId)(dispatch);
     });
 
     resolve();