+++ src/containers/stages/settings/SettingsContainer.js
@@ -3,13 +3,16 @@ import Settings from '../../../components/stages/settings/Settings';
 import {MODULE_ID_SETTINGS} from '../../../constants';
 import translations from '../../../configs/translations';
 import getTranslation from '../../../helpers/getTranslation';
+import { bindActionCreators } from 'redux';
+import saveUserSettings from '../../../actions/settings/saveUserSettings';
 
 const mapStateToProps = (state) => {
+
     const displayLanguages = Object.keys(translations).map(locale => {
         const languageLabel = getTranslation(state, `LANGUAGE_${locale.replace('-','_').toUpperCase()}`);
 
         if (!languageLabel) {
-            return;
+            return undefined;
         }
 
         return {
@@ -24,9 +27,15 @@ const mapStateToProps = (state) => {
         renderModule: MODULE_ID_SETTINGS === state.activeModule,
         translations: {
             displayLanguageLabel: getTranslation(state, 'SETTINGS_DISPLAY_LANGUAGE'),
+            pageTitle: getTranslation(state, 'PAGE_TITLE'),
+            saveButtonLabel: getTranslation(state, 'SETTINGS_SAVE_BUTTON_LABEL'),
             titleLabel: getTranslation(state, 'SETTINGS_TITLE'),
         }
     };
 };
 
-export default connect(mapStateToProps)(Settings);
+const mapDispatchToProps = (dispatch) => {
+    return bindActionCreators({ saveUserSettings }, dispatch);
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(Settings);